Refactor services => service, move MarkdownDocument into service
